# ext/opcache: Retry waitpid on EINTR from repeated reload signals